+++ lib/mailchimp_api/resources/list.rb
@@ -5,5 +5,7 @@ module MailchimpAPI
     extend MailchimpAPI::Support::Countable
 
     self.collection_parser = CollectionParsers::List
+
+    has_many :interest_categories, class_name: 'MailchimpAPI::InterestCategory'
   end
 end